+++ extension/src/register_types.cpp
@@ -1,0 +1,35 @@
+#include "webtransport.hpp"
+
+#include <godot_cpp/core/class_db.hpp>
+#include <godot_cpp/godot.hpp>
+
+using namespace godot;
+
+static void initialize_godot_wtransport(ModuleInitializationLevel p_level) {
+    if (p_level != MODULE_INITIALIZATION_LEVEL_SCENE) {
+        return;
+    }
+    GDREGISTER_CLASS(WebTransportTlsOptions);
+    GDREGISTER_CLASS(WebTransportStream);
+    GDREGISTER_CLASS(WebTransportSession);
+    GDREGISTER_CLASS(WebTransportClient);
+}
+
+static void uninitialize_godot_wtransport(ModuleInitializationLevel p_level) {
+    if (p_level != MODULE_INITIALIZATION_LEVEL_SCENE) {
+        return;
+    }
+}
+
+extern "C" {
+GDExtensionBool GDE_EXPORT godot_wtransport_library_init(
+        GDExtensionInterfaceGetProcAddress p_get_proc_address,
+        GDExtensionClassLibraryPtr p_library,
+        GDExtensionInitialization *r_initialization) {
+    GDExtensionBinding::InitObject init_object(p_get_proc_address, p_library, r_initialization);
+    init_object.register_initializer(initialize_godot_wtransport);
+    init_object.register_terminator(uninitialize_godot_wtransport);
+    init_object.set_minimum_library_initialization_level(MODULE_INITIALIZATION_LEVEL_SCENE);
+    return init_object.init();
+}
+}